Move AssetValidator to util package
