+++ core/src/main/java/org/stellar/anchor/sep24/Sep24Service.java
@@ -5,10 +5,10 @@ import static org.stellar.anchor.api.event.AnchorEvent.Type.TRANSACTION_CREATED;
 import static org.stellar.anchor.api.sep.SepTransactionStatus.INCOMPLETE;
 import static org.stellar.anchor.api.sep.sep24.InfoResponse.FeatureFlagResponse;
 import static org.stellar.anchor.api.sep.sep24.InfoResponse.FeeResponse;
-import static org.stellar.anchor.asset.AssetServiceValidator.*;
 import static org.stellar.anchor.event.EventService.EventQueue.TRANSACTION;
 import static org.stellar.anchor.sep24.Sep24Helper.fromTxn;
 import static org.stellar.anchor.sep24.Sep24Transaction.Kind.*;
+import static org.stellar.anchor.util.AssetValidator.*;
 import static org.stellar.anchor.util.Log.debug;
 import static org.stellar.anchor.util.Log.debugF;
 import static org.stellar.anchor.util.Log.info;
